+++ src/wai/annotations/converters/tfexample/_ToTensorflowExample.py
@@ -7,10 +7,10 @@ from wai.common.geometry import Polygon
 from wai.common.adams.imaging.locateobjects import LocatedObjects
 import tensorflow as tf
 
-from ...core import InternalFormatConverter, ImageFormat, ImageInfo
+from ...core import InternalFormatConverter, ImageInfo
 from ...core.constants import LABEL_METADATA_KEY
 from ...core.external_formats import TensorflowExampleExternalFormat
-from ...tf_utils import make_feature
+from ...tf_utils import make_feature, negative_example
 
 
 class ToTensorflowExample(InternalFormatConverter[TensorflowExampleExternalFormat]):
@@ -32,12 +32,9 @@ class ToTensorflowExample(InternalFormatConverter[TensorflowExampleExternalForma
         if image_info.data is None:
             raise ValueError(f"Tensorflow records require image data")
 
-        # Get the image format
-        image_format = ImageFormat.for_filename(image_info.filename)
-
         # If no annotations, return an empty example
         if len(located_objects) == 0:
-            return self.create_empty_example(image_info, image_format)
+            return negative_example(image_info)
 
         # Format and extract the relevant annotation parameters
         lefts, rights, tops, bottoms, labels, classes, masks = self.process_located_objects(located_objects,
@@ -51,7 +48,7 @@ class ToTensorflowExample(InternalFormatConverter[TensorflowExampleExternalForma
             'image/filename': make_feature(image_info.filename),
             'image/source_id': make_feature(image_info.filename),
             'image/encoded': make_feature(image_info.data),
-            'image/format': make_feature(image_format.get_default_extension()),
+            'image/format': make_feature(image_info.format.get_default_extension()),
             'image/object/bbox/xmin': make_feature(lefts),
             'image/object/bbox/xmax': make_feature(rights),
             'image/object/bbox/ymin': make_feature(tops),
@@ -139,32 +136,6 @@ class ToTensorflowExample(InternalFormatConverter[TensorflowExampleExternalForma
                                                         image_height))
 
         return lefts, rights, tops, bottoms, labels, classes, masks
-
-    def create_empty_example(self, image_info: ImageInfo, image_format: ImageFormat):
-        """
-        Creates an empty example (for images with no annotations).
-
-        :return:    The empty example.
-        """
-        return tf.train.Example(
-            features=tf.train.Features(
-                feature={
-                    'image/height': make_feature(image_info.height()),
-                    'image/width': make_feature(image_info.width()),
-                    'image/filename': make_feature(image_info.filename),
-                    'image/source_id': make_feature(image_info.filename),
-                    'image/encoded': make_feature(image_info.data),
-                    'image/format': make_feature(image_format.get_default_extension()),
-                    'image/object/bbox/xmin': tf.train.Feature(float_list=tf.train.FloatList(value=[])),
-                    'image/object/bbox/xmax': tf.train.Feature(float_list=tf.train.FloatList(value=[])),
-                    'image/object/bbox/ymin': tf.train.Feature(float_list=tf.train.FloatList(value=[])),
-                    'image/object/bbox/ymax': tf.train.Feature(float_list=tf.train.FloatList(value=[])),
-                    'image/object/class/text': tf.train.Feature(bytes_list=tf.train.BytesList(value=[])),
-                    'image/object/class/label': tf.train.Feature(int64_list=tf.train.Int64List(value=[])),
-                    'image/object/mask': tf.train.Feature(bytes_list=tf.train.BytesList(value=[]))
-                }
-            )
-        )
 
     def mask_from_polygon(self, polygon: Polygon, image_width: int, image_height: int) -> bytes:
         """