+++ src/wai/annotations/io/tfexample/_TensorflowExampleReader.py
@@ -1,32 +1,27 @@
-from typing import Iterable, Iterator
+from typing import Iterator
 
 import tensorflow as tf
 
-from ...core import Reader
+from ...core import Reader, ImageInfo
 from ...core.external_formats import TensorflowExampleExternalFormat
-from ...tf_utils import sharded_filename_params, format_sharded_filename
+from ...tf_utils import ShardCompactor, is_shard_filename, get_all_sharded_filenames, negative_example
 
 
 class TensorflowExampleReader(Reader[TensorflowExampleExternalFormat]):
     """
     Reader of Tensorflow Example records.
     """
-    def determine_input_files(self, input_path: str) -> Iterable[str]:
-        # Treat the file as a shard file
-        base_path, shards, index = sharded_filename_params(input_path)
+    def annotation_files(self) -> Iterator[str]:
+        # Make sure each set of sharded files only appears once
+        return ShardCompactor().process(super().annotation_files())
 
-        # If it's not a sharded filename, return the path itself
-        if shards == 0:
-            return input_path,
+    def read_annotation_file(self, filename: str) -> Iterator[TensorflowExampleExternalFormat]:
+        # Recurse if given a shard filename
+        if is_shard_filename(filename):
+            for shard_filename in get_all_sharded_filenames(filename):
+                yield self.read_annotation_file(shard_filename)
+            return
 
-        # If it is a sharded filename, iterate over all matching shards
-        return (format_sharded_filename(base_path, shards, index) for index in range(shards))
-
-    @classmethod
-    def input_help_text(cls) -> str:
-        return "tensorflow records file (if sharded, any of the shard files)"
-
-    def read(self, filename: str) -> Iterator[TensorflowExampleExternalFormat]:
         # Load the record file
         dataset = tf.data.TFRecordDataset([filename])
 
@@ -39,3 +34,6 @@ class TensorflowExampleReader(Reader[TensorflowExampleExternalFormat]):
             instance.ParseFromString(record.numpy())
 
             yield instance
+
+    def image_info_to_external_format(self, image_info: ImageInfo) -> TensorflowExampleExternalFormat:
+        return negative_example(image_info)